requirements.txt
celery==4.4
Cython==0.29.21
h5py==2.10.0
hdbscan==0.8.23
keras==2.2.4
Markdown==3.2.2
matplotlib==3.3.0
networkx==2.4
numpy==1.19.0
pandas==1.0.5
pillow==7.2.0
pycairo==1.19.1
pymongo==3.10.1
pytorch>=1.4.0
PyYAML==5.3.1
rdchiral==1.0.0
scipy==1.5.1
scikit-learn==0.23.1
sphinx==3.1.2
tensorflow==2.0.0
tqdm==4.48.0
CairoSVG==2.4.2
